app/src/main/java/example/codeclan/com/movieranking/Movie.java
package example.codeclan.com.movieranking;

/**
 * Created by user on 16/01/2017.
 */

public class Movie {

    private String title;
    private String genre;
    private int ranking;

    public Movie(String title, String genre, int ranking) {
        this.title = title;
        this.genre = genre;
        this.ranking = ranking;
    }

    public String getTitle() {
        return this.title;
    }

    public String getGenre() {
        return this.genre;
    }

    public int getRanking() {
        return this.ranking;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public void setGenre(String genre) {
        this.genre = genre;
    }

    public void setRanking(int ranking) {
        this.ranking = ranking;
    }

    @Override
    public String toString() {
        return "Title: '" + title + '\'' +
                ", Genre: '" + genre + '\'' +
                ", Ranking: " + ranking;
    }

}
